% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_inat.R
\name{assign_inat}
\alias{assign_inat}
\title{Assign iNaturalist observations to FWS property}
\usage{
assign_inat(fwsinat, progress = TRUE)
}
\arguments{
\item{fwsinat}{\code{data.frame} of iNaturalist observations produced by running
\code{\link{retrieve_inat}}}

\item{progress}{logical (default = TRUE) should progress bar be displayed when
assigning records to the nearest USFWS property}
}
\value{
\code{fwsinat} \code{data.frame} of iNaturalist observations with the USFWS
 property (i.e., \code{orgname}) appended as the first column
}
\description{
Assignment occurs in a 2-stage process.  First, the locational coordinates of the
 observations are compared directly with USFWS property boundaries.  Second, points
 not falling within a boundary *and* identified as having obscured locations are assigned
 to the nearest property within ~ 50 km, if any.
}
\examples{
\dontrun{
fws <- retrieve_inat()
fws <- assign_inat(fws)
}
}
