% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_rows.R
\name{bind_rows}
\alias{bind_rows}
\title{Use dplyr::bind_rows on fwsinat objects}
\usage{
bind_rows(...)
}
\arguments{
\item{...}{Inputs passed to dplyr::bind_rows}
}
\value{
If all inputs are \code{fwsinat} objects, then the attributes
 are stacked and an \code{fwsinat} will be returned.  Otherwise, the expected
 outcome of \code{\link[dplyr]{left_join}} is returned and a message is displayed.
}
\description{
Since \code{dplyr::bind_rows} is not currently an S3 method, this function
 masks \code{dplyr::bind_rows} (although it is called directly) for use on
 \code{fwsinat} objects.
}
\seealso{
\code{\link[dplyr]{bind_rows}}

\url{https://github.com/tidyverse/dplyr/issues/2457}
}
