% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_inat.R
\name{export_inat}
\alias{export_inat}
\title{Export iNaturalist observations to Excel file(s) based on for assigned to FWS properties  refuges}
\usage{
export_inat(fwsinat, dir = NULL, xl_out, overwrite = TRUE, verbose = TRUE)
}
\arguments{
\item{fwsinat}{\code{fwsinat} \code{data.frame} of iNaturalist observations
produced by running \code{\link{retrieve_inat}}, and (optionally) subsequently through
\code{\link{assign_inat}}}

\item{dir}{a non-empty character scalar giving the directory within which to store
exported spreadsheets}

\item{xl_out}{character string indicating the output file name (without extension);
note that periods and commas will be removed and spaces will be replaced with underscores}

\item{overwrite}{logical (default \code{TRUE}) indicating whether to overwrite an existing file}

\item{verbose}{logical (default \code{TRUE}) indicating whether to provide messaging during
export process}
}
\value{
\code{NULL}; Exports Excel file(s) to \code{dir}
}
\description{
Export iNaturalist observations to Excel file(s) based on for assigned to FWS properties  refuges
}
\examples{
\dontrun{
fws <- retrieve_inat()
export_inat(fws, "./fws.xlsx")
}
}
