% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_inat.R
\name{harvest_inat}
\alias{harvest_inat}
\title{Harvest observations outside the USFWS National Wildlife Refuge System
iNaturalist project for a given property}
\usage{
harvest_inat(refuge = NULL, user = NULL, pw = NULL, interactive = TRUE)
}
\arguments{
\item{refuge}{character string or vector indicating USFWS properties
on which to harvest observations not associated with the USFWS
National Wildlife Refuge System project ("usfws-national-wildlife-refuge-system";
\url{http://www.inaturalist.org/projects/usfws-national-wildlife-refuge-system}).
It is strongly advised that this parameter result from running
\code{\link{find_refuges}} avoid potential mismatches.  See Examples. By
default (\code{refuge = NULL}), all USFWS properties with boundaries
in iNaturalist are processed sequentially.}

\item{user}{character scalar of iNaturalist username or associated e-mail address}

\item{pw}{character scalar of iNaturalist password}

\item{interactive}{if \code{TRUE}, the default, the user is prompted to proceed
prior to completing the harvest.  Setting to \code{FALSE} proceeds with the
harvest without prompting}
}
\value{
This function performs a series of POST operations to assign observations
 to the USFWS NWRS project via the iNaturalist API (\url{???}). As a side effect,
 it returns a \code{data.frame} of observations that failed to be assigned, if any.
}
\description{
Harvest observations outside the USFWS National Wildlife Refuge System
iNaturalist project for a given property
}
\examples{
\dontrun{
### CREATE A MEANINGFUL EXAMPLE
}
}
