% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inat_export.R
\name{inat_export}
\alias{inat_export}
\title{Export iNaturalist observations to Excel file(s) based on for assigned to FWS properties  refuges}
\usage{
inat_export(fwsinat, dir = NULL, overwrite = TRUE, verbose = TRUE)
}
\arguments{
\item{fwsinat}{\code{fwsinat} \code{data.frame} of iNaturalist observations
produced by running \code{\link{inat_retrieve}}}

\item{dir}{a non-empty character scalar giving the directory within which to store
exported spreadsheets}

\item{overwrite}{logical (default \code{TRUE}) indicating whether to overwrite an existing file}

\item{verbose}{logical (default \code{TRUE}) indicating whether to provide messaging during
export process}
}
\value{
\code{NULL}; Exports Excel file(s) to \code{dir}
}
\description{
Export iNaturalist observations to Excel file(s) based on for assigned to FWS properties  refuges
}
\examples{
\dontrun{
fws <- inat_retrieve()
inat_export(fws, "./fws.xlsx")
}
}
