% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inat_retrieve.R
\name{inat_retrieve}
\alias{inat_retrieve}
\title{Retrieve iNaturalist observations on USFWS properties}
\usage{
inat_retrieve(refuge = NULL,
  inat_proj = "usfws-national-wildlife-refuge-system", taxon_name = NULL,
  d1 = NULL, d2 = NULL, since_date = NULL, multipart = FALSE,
  verbose = TRUE)
}
\arguments{
\item{refuge}{character scalar or vector indicating USFWS properties
on which to retrieve iNaturalist observations.  It is strongly advised that
the input for this parameter result from running \code{\link{find_refuges}} to
avoid potential mismatches.  See examples. By default, all 500+ USFWS
properties with boundaries in iNaturalist are used.}

\item{inat_proj}{character scalar of an iNaturalist project as an iNaturalist
slug or group ID. Default is to retrieve all records on the property associated
with the USFWS National Wildlife Refuge System project
("usfws-national-wildlife-refuge-system";
\url{http://www.inaturalist.org/projects/usfws-national-wildlife-refuge-system}).
Another potentially useful option is to retrieve all available iNaturalist records
on the property using (\code{inat_proj = NULL})}

\item{taxon_name}{character scalar. Retrieve only observations associated with
\code{taxon_name}. This also retrieves descendant taxa. Note if \code{taxon_name} is
ambiguous and matches multiple taxa, no observations may be retrieved.}

\item{d1}{First date of a date range as string in the form yyyy-mm-dd (e.g. "2017-07-01").
Specifying only \code{d1} returns all records on and after \code{d1}.}

\item{d2}{Last date of a date range as string in the form yyyy-mm-dd (e.g. "2017-07-31").
Specifying only \code{d2} returns all records on and before \code{d2}.}

\item{since_date}{string with ISO 8601 datetime (e.g. "2017-07-04" or
"2017-07-04T13:40:13-05:00"). Only records added or updated on iNaturalist after
this date will be returned. Primarily intended for internals use from
\code{\link{inat_update}}.}

\item{multipart}{logical (default \code{FALSE}) indicating whether the current retrieval
is part of multiple retrievals to obtain > 10000 records.  If \code{TRUE}, the current
system time is recorded as the the \code{query_dt} attribute of the resulting
\code{fwsinat} object rather than \code{d2}, if specified.}

\item{verbose}{logical (default \code{TRUE}); print informative messages?}
}
\value{
\code{fwsinat} object (essentially a \code{data.frame} with a few additional
 attributes) of iNaturalist observations associated with the supplied \code{refuge}(s)
 and \code{inat_proj}, if any, and joined, when possible, with ITIS
 (\url{http://www.itis.gov}) information
}
\description{
The observations can also be restricted to a single iNaturalist project,
 which is associated with a "slug", usually in the form of the project name
 as a single lower case string with words seperated by hyphens. For instance,
 the default "USFWS National Wildlife Refuge System" project has a slug of
 "usfws-national-wildlife-refuge-system". The slug can be extracted from the
 for the project. The USFWS NWRS project, for example, resides at
 \url{http://www.inaturalist.org/projects/usfws-national-wildlife-refuge-system}
}
\examples{
\dontrun{
# Default is to retrieve records for all available USFWS properties from the USFWS
# National Wildlife Refuge System projects
fws <- inat_retrieve()
fws_july17 <- inat_retrieve(d1 = "2017-07-01", d2 = "2017-07-31")

# But it works with other projects too
bon <- find_refuges("bon secour")
bs <- inat_retrieve(bon, "bon-secour-national-wildlife-refuge-bioblitz")

# Or get all available iNaturalist observations on the property
bon <- find_refuges("bon secour")
bs_all <- inat_retrieve(bon, inat_proj = NULL)
}
}
\seealso{
\code{\link{inat_update}} for updating a previous iNaturalist data retrieval

\code{\link{inat_export}} for exporting an iNaturalist data retrieval
}
