% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_inat.R
\name{retrieve_inat}
\alias{retrieve_inat}
\title{Retrieve observations from an iNaturalist project and link with ITIS}
\usage{
retrieve_inat(inat_proj = "usfws-national-wildlife-refuge-system",
  d1 = NULL, d2 = NULL, since_date = NULL)
}
\arguments{
\item{inat_proj}{name of the iNaturalist project as an iNaturalist slug or group ID.
Default is to retrieve records for the USFWS National Wildlife Refuge project
(\url{http://www.inaturalist.org/projects/usfws-national-wildlife-refuge-system}).}

\item{d1}{First date of a date range as string in the form yyyy-mm-dd (e.g. "2017-07-01").
Specifying only \code{d1} returns all records on and after \code{d1}.}

\item{d2}{Last date of a date range as string in the form yyyy-mm-dd (e.g. "2017-07-31").
Specifying only \code{d2} returns all records on and before \code{d2}.}

\item{since_date}{string with ISO 8601 datetime (e.g. "2017-07-04" or
"2017-07-04T13:40:13-05:00"). Only records added or updated on iNaturalist after
this date will be returned. Primarily intended to be used internally from
\code{\link{update.fwsinat}}.}
}
\value{
\code{fwsinat} \code{data.frame} of iNaturalist observations associated with the
 \code{inat_proj} joined, when possible, with ITIS (\url{http://www.itis.gov})
 information
}
\description{
An iNaturalist slug is usually the project name as single lower case
 string with words seperated by hyphens. For instance, the project "USFWS
 National Wildlife Refuge System" has a slug of
 "usfws-national-wildlife-refuge-system". The slug can also be extracted
 from the URL for the project. The USFWS NWRS project, for example, resides at
 \url{http://www.inaturalist.org/projects/usfws-national-wildlife-refuge-system}
}
\examples{
\dontrun{
# Default is to retrieve records for USFWS National Wildlife Refuge System project
fws <- retrieve_inat()
fws_july17 <- retrieve_inat(d1 = "2017-07-01", d2 = "2017-07-31")

# But it works with other projects too
bs <- retrieve_inat("bon-secour-national-wildlife-refuge-bioblitz")
}
}
\seealso{
\code{\link{bind_rows}} for joining multiple iNaturalist data retrievals while
 preserving attributes

\code{\link{assign_inat}} for assigning iNaturalist observations to FWS
 properties

\code{\link{update.fwsinat}} for updating a previous iNaturalist data retrieval

\code{\link{export_inat}} for exporting a previous iNaturalist data retrieval
}
