% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fwsinat}
\alias{summary.fwsinat}
\title{fwsinat summary}
\usage{
\method{summary}{fwsinat}(object, ...)
}
\arguments{
\item{object}{\code{fwsinat} \code{data.frame} of iNaturalist observations
produced by running \code{\link{retrieve_inat}}, and (optionally) subsequently through
\code{\link{assign_inat}} and/or \code{\link{update.fwsinat}}}

\item{...}{see \link[base]{summary}}
}
\value{
Same as \link[base]{summary.data.frame} but also prints information on contained
 iNaturalist projects and the date of the retrieval.
}
\description{
fwsinat summary
}
