% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.fwsinat}
\alias{update.fwsinat}
\title{Update previous retrieval of observations from an iNaturalist project}
\usage{
\method{update}{fwsinat}(object, ...)
}
\arguments{
\item{object}{\code{fwsinat} \code{data.frame} of iNaturalist observations
produced by running \code{\link{retrieve_inat}}, and (optionally) subsequently through
\code{\link{assign_inat}}}

\item{...}{not used in this implementation}
}
\value{
\code{fwsinat} \code{data.frame} of iNaturalist observations associated with the
 \code{inat_proj} joined, when possible, with ITIS (\url{http://www.itis.gov})
 information
}
\description{
This function uses the query timestamp associated with the \code{fwsinat}
 \code{object} to perform a new
}
\examples{
\dontrun{
# Default is to retrieve records for USFWS National Wildlife Refuge System project
fws <- retrieve_inat()

# Wait a while... hours, days, months...whatever
fws <- update(fws)
}
}
