% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_inat}
\alias{update_inat}
\title{Update previous retrieval of observations from an iNaturalist project}
\usage{
update_inat(fwsinat)
}
\arguments{
\item{fwsinat}{\code{fwsinat} \code{data.frame} of iNaturalist observations
produced by running \code{\link{retrieve_inat}}}
}
\value{
\code{fwsinat} object of iNaturalist observations associated
 with one or more USFWS properties, and potentially an iNaturalist project.
 These observations are joined, when possible, with ITIS
 (\url{http://www.itis.gov}) information.
}
\description{
Uses the query timestamp associated with the \code{fwsinat}
 object to perform a new iNaturalist retrieval for the associated
 USFWS properties (and possibly project) and updates records that have
 changed and adds records created since the previous query.
}
\examples{
\dontrun{
# Default is to retrieve observations from the USFWS National Wildlife
# Refuge System project on all available USFWS properties
fws <- retrieve_inat()

# Wait a while... hours, days, months... for users to update and add new
# observations
fws <- update_inat(fws)
}
}
\seealso{
\code{\link{retrieve_inat}} for details on making the initial
 retrieval of iNaturalist observations
}
