% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_rows.R
\name{bind_rows}
\alias{bind_rows}
\title{Bind Rows (fwsinat)}
\usage{
bind_rows(...)
}
\arguments{
\item{...}{Inputs passed to dplyr::bind_rows}
}
\value{
If all inputs are `fwsinat` objects, then the attributes
 are stacked and an `fwsinat` will be returned.  Otherwise, the expected
 outcome of `dplyr::bind_rows` is returned and a message is displayed.
}
\description{
Since dplyr::bind_rows is not currently an S3 method, this function
 masks dplyr::bind_rows (although it is called directly) for use on
 `fwsinat`` objects.
}
\seealso{
[Related dplyr issue](https://github.com/tidyverse/dplyr/issues/2457)

\code{\link[dplyr]{bind_rows}}
}
