% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_refuges.R
\name{find_refuges}
\alias{find_refuges}
\title{Find refuges or properties available for query.}
\usage{
find_refuges(refuge = NULL, ptype = c("NWR", "WMD", "WMA", "WPA"),
  region = 1:8L)
}
\arguments{
\item{refuge}{character string scalar or vector (i.e., multiple entries allowed and
regular expressions allowed) with
which to search and return valid refuge identifiers for use with
\code{\link{retrieve_inat}}. Default (`NULL`) returns all available refuges.
Case-insensitive search. Regular expressions allowed. See Examples.}

\item{ptype}{character string scalar or vector of types of USFWS properties to search.
Default is to search all property types.  Other viable options include National Wildlife
Refuges only (`ptype = "NWR"`), "WMD" (Wetland Managment District), "WMA" (Wildlife
Management Area), and WPA" (Waterfowl Production Area).  Multiple options are
permissible.}

\item{region}{integer vector indicating which USFWS Region to search; see
\url{https://www.fws.gov/where}; valid values range from 1 to 8}
}
\value{
character scalar or vector of proper organizational name (ORGNAME) for USFWS
 properties meeting the search criteria.  This output can be passed directly as the
 \code{refuge} argument in \code{\link{retrieve_inat}}.
}
\description{
Find refuges or properties available for query.
}
\examples{
\dontrun{
# Get all available properties
all_props <- find_refuges()

# Get all National Wildlife Refuges
all_refs <- find_refuges(ptype = "NWR")

# Search for refuges using a partial name match
ml <- find_refuges("longleaf")

# Search for refuges matching multiple strings
multi <- find_refuges(c("longleaf", "romain"))

# Same search using regular expressions
multi <- find_refuges("longleaf|romain")

# Return all southeast (region 4) refuges
r4 <- find_refuges(region = 4, ptype = "NWR")
}
}
