% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.fwsinat}
\alias{update.fwsinat}
\title{Update previous retrieval of observations from an iNaturalist project}
\usage{
\method{update}{fwsinat}(object, ...)
}
\arguments{
\item{object}{\code{fwsinat} object of iNaturalist observations
produced by \code{\link{retrieve_inat}}}

\item{...}{not used in this implementation}
}
\value{
\code{fwsinat} object of iNaturalist observations associated
 with one or more USFWS properties, and potentially an iNaturalist project.
 These observations are joined, when possible, with ITIS
 (\url{http://www.itis.gov}) information.
}
\description{
Uses the query timestamp associated with the \code{fwsinat}
 \code{object} to perform a new iNaturalist retrieval for the associated
 USFWS properties (and possibly project) and updates records that have
 changed and adds records created since the previous query.
}
\examples{
\dontrun{
# Default is to retrieve observations from the USFWS National Wildlife
# Refuge System project on all available USFWS properties
fws <- retrieve_inat()

# Wait a while... hours, days, months... for users to update and add new
# observations
fws <- update(fws)
}
}
\seealso{
\code{\link{retrieve_inat}} for details on making the initial
 retrieval of iNaturalist observations
}
